Drop failing extract-win-cert test.
This used to test that trying to load a cert with Windows line endings
will yield an error. Recent OpenSSL has become more lenient in parsing
so this will now successfully verify (rightly so I guess). So the test
fails when using OpenSSL 1.1 but is then also rather obsolete.

Because this really only tests the OpenSSL library and not something
specific to xmlseclibs, I think its best to just drop it.

Fixes #118