Make SignatureAlgorithmFactory reusable for custom algs
This commit rewrites the way the signature algorithms work and are created via the factory. It offloads the responsibility to check what algorithms they support to the signer implementations themselves, and makes them return their supported algs.

It also introduces a mechanism in the factory so that you can register new signer implementations, and use them automatically in your code.

Signed-off-by: Jaime Pérez <jaimepc@gmail.com>

Update tests for SignatureAlgorithmFactory