First attempt to implement signing in XML objects
This introduces a SignedElementTrait that gives us most of what we need.

The SignableElementInterface::sign() method is now changed to receive a SignatureAlgorithm, the canonicalization algorithm and an optional KeyInfo element to add to the resulting signature. It also introduces a new method getId() that each element must implement to return their own ID used by the signature reference.

The new trait implements the sign() method of the interface, and also provides a doSign() method that objects can use to effectively sign them. Calling it will generate a Signature object and keep it in a property. This object can then be used in toXML() in the object and inserted wherever we want, using the insertBefore() convenience method (also provided by the trait).

The CustomSignable class demonstrates how to use the functionality. CustomSigned and its corresponding test disappears (as it made no sense in this context, since the actual XML elements are called CustomSignable).

Signed-off-by: Jaime Pérez <jaimepc@gmail.com>