Signable elements should call toXML()
We got rid of the initial extra toSignedXML() method. If we want to sign an element, we should just call sign() on it, then get the result with toXML().

This means it will be toXML() the method that actually does the signing, while sign() will simply prepare everything for it and signal that the element should be signed when calling toXML().

Signed-off-by: Jaime Pérez <jaimepc@gmail.com>