Implement signature verification, refactor signing
This commit adds support for verifying signatures (and renames the API method to verify(), which is the correct term we should us in this context, instead of validate()).

It also refactors the way we sign, and adds support for canonicalization algorithms with comments, which were in practice not supported before.

Signed-off-by: Jaime Pérez <jaimepc@gmail.com>